Emscripten: Fixes for Iup official resync. iupdrvFontGetTextSize is a hack and needs to be reviewed.

+++ src/emscripten/iupemscripten_info.c
@@ -60,7 +60,7 @@ int iupdrvGetScreenDepth(void)
 	return 32;
 }
 
-float iupdrvGetScreenDpi(void)
+double iupdrvGetScreenDpi(void)
 {
 #if 0
 	CGRect rect = CGDisplayBounds(kCGDirectMainDisplay);
